+++ src/tdl/tdl.h
@@ -1,13 +1,13 @@
 // -----------------------------------------------------------------------------------------------------
-// Copyright (c) 2006-2022, Knut Reinert & Freie Universität Berlin
-// Copyright (c) 2016-2022, Knut Reinert & MPI für molekulare Genetik
+// Copyright (c) 2006-2023, Knut Reinert & Freie Universität Berlin
+// Copyright (c) 2016-2023, Knut Reinert & MPI für molekulare Genetik
 // This file may be used, modified and/or redistributed under the terms of the 3-clause BSD-License
 // shipped with this file.
 // -----------------------------------------------------------------------------------------------------
 
 #pragma once
 
-#include "ParamDocument.h"
-#include "ParamDocumentToCTD.h"
-#include "ParamDocumentToCWL.h"
+#include "ToolInfo.h"
+#include "convertToCTD.h"
+#include "convertToCWL.h"
