+++ src/test_tdl/multiple_definition01_test.cpp
@@ -1,12 +1,12 @@
 // -----------------------------------------------------------------------------------------------------
-// Copyright (c) 2006-2022, Knut Reinert & Freie Universität Berlin
-// Copyright (c) 2016-2022, Knut Reinert & MPI für molekulare Genetik
+// Copyright (c) 2006-2023, Knut Reinert & Freie Universität Berlin
+// Copyright (c) 2016-2023, Knut Reinert & MPI für molekulare Genetik
 // This file may be used, modified and/or redistributed under the terms of the 3-clause BSD-License
 // shipped with this file.
 // -----------------------------------------------------------------------------------------------------
 
-#include <tdl/ParamDocument.h>
-#include <tdl/ParamDocumentToCTD.h>
+#include <tdl/ToolInfo.h>
+#include <tdl/convertToCTD.h>
 
 
 /* This file, in combination with multiple_definition02_test.cpp are testing for multiple definition errors.
